AD-HOC fix (configuration): Reload configuration on change
Currently there is a problem in which the configuration for the
`td-agent-bit` daemon will not be reloaded when it is changed. This
means for the *first* ansible task all things will appear to be correct,
but for *subsequent* ansible tasks the configuration will not apply to
the daemon.

This commit introduces reloading of the daemon dynamically based on
whether this configuration has changed, using ansibles "handler"
primitives.